# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CustomPlan } from './customPlan';
import type { Order } from './order';
import type { User } from './user';

export interface Account {
  user: User;
  orders: Order[];
  customPlans: CustomPlan[];
  savedTourIds: number[];
}