# lib/api-zod/src/generated/types/adminSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from './order';

export interface AdminSummary {
  totalUsers: number;
  totalTours: number;
  totalOrders: number;
  totalPlans: number;
  revenue: number;
  recentOrders: Order[];
}