# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  name: string;
  email: string;
  phone: string;
  subject: string;
  /** @minLength 5 */
  message: string;
}