# lib/api-zod/src/generated/types/customPlan.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ItineraryDay } from './itineraryDay';

export interface CustomPlan {
  id: number;
  destinations: string[];
  startDate: string;
  endDate: string;
  duration: number;
  travelers: number;
  style: string;
  interests: string[];
  accommodation: string;
  transportation: string;
  estimatedBudget: number;
  itinerary: ItineraryDay[];
  status: string;
}