# lib/api-zod/src/generated/types/destination.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Destination {
  id: number;
  name: string;
  state: string;
  region: string;
  description: string;
  image: string;
  attractions: string[];
  startingPrice: number;
}